Changed the data type

+++ python_model/gen_dataset.py
@@ -21,18 +21,18 @@ def gen_data():
 	path = os.path.abspath(os.getcwd()) + "/data/reinforce1/001"
 
 	struct1og, vgc, vGextC, vGextF, vGstayOff = eng.get_struct1(nargout=5)
-	struct1 = np.int8(np.array(struct1og))
+	struct1 = np.array(struct1og, dtype=np.int8)
 	(x,y,z) = struct1.shape
-	struct1C = np.array(vGextC)
-	struct1F = np.array(vGextF)
-	struct1Off = np.array(vGstayOff)
+	struct1C = np.array(vGextC, dtype=np.int8)
+	struct1F = np.array(vGextF, dtype=np.int8)
+	struct1Off = np.array(vGstayOff, dtype=np.int8)
 	datasize = 20
 	structs = []
 	rein = []
 	for i in range(datasize):
 		print(i)
-		true_struct = np.int8(np.array(eng.reinforce_struct(matlab.int8(struct1.tolist()), vGextC, 
-													vGextF, vGstayOff, 200)))
+		true_struct = np.array(eng.reinforce_struct(matlab.int8(struct1.tolist()), vGextC, 
+													vGextF, vGstayOff, 200), dtype=np.int8)
 		
 		structs.append(struct1)
 		rein.append(true_struct)
